user wallet, improvements, proxy io integration

+++ app/Http/Controllers/Api/AuthController.php
@@ -4,18 +4,22 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\AuthRequest;
+use App\Http\Resources\FailedResource;
 use App\Http\Resources\UserResource;
 use Illuminate\Http\JsonResponse;
 use App\Services\UserService;
-use Illuminate\Support\Facades\Log;
 
 
 class AuthController extends Controller
 {
 
-    public function logout(): JsonResponse
+    public function logout(): JsonResponse|FailedResource
     {
         $user = auth('sanctum')->user();
+
+        if (!$user) {
+            return new FailedResource(401, 'No authenticated user found.');
+        }
 
         try {
             $user->currentAccessToken()->delete();
@@ -23,59 +27,37 @@ class AuthController extends Controller
                 'status' => 'success',
                 'message' => 'Successfully logged out',
             ]);
-        } catch (\Exception $e) {
-            Log::error('Logout Failed', ['error' => $e->getMessage()]);
-
-            return response()->json([
-                'status' => 'error',
-                'message' => 'Failed to logout, please try again.',
-            ], 500);
+        } catch (\Exception) {
+            return new FailedResource(500, 'Failed to logout, please try again');
         }
     }
 
-    public function login(AuthRequest $request, UserService $userService): JsonResponse
+    public function login(AuthRequest $request, UserService $userService): UserResource|FailedResource
     {
         try {
             $credentials = $request->only('email', 'password');
             $user = $userService->attemptLogin($credentials, $request->device_name);
 
             if ($user) {
-                return response()->json([
-                    'status' => 'success',
-                    'user' => new UserResource($user),
-                ]);
+                return new UserResource($user);
             } else {
-                return response()->json([
-                    'status' => 'error',
-                    'message' => 'Invalid login credentials.',
-                ], 401);
+                return new FailedResource(401, 'Invalid login credentials.');
             }
-        } catch (\Exception $e) {
-            Log::error('Login Failed', ['error' => $e->getMessage()]);
 
-            return response()->json([
-                'status' => 'error',
-                'message' => 'An error occurred during login. Please try again later.',
-            ], 500);
+        } catch (\Exception) {
+            return new FailedResource(500, 'An error occurred during login. Please try again later.');
         }
     }
 
 
-    public function register(AuthRequest $request, UserService $userService): JsonResponse
+    public function register(AuthRequest $request, UserService $userService): UserResource|FailedResource
     {
         try {
             $user = $userService->createUser($request->all());
-            return response()->json([
-                'status' => 'success',
-                'user' => new UserResource($user),
-            ]);
-        } catch (\Exception $e) {
-            Log::error('User Registration Failed', ['error' => $e->getMessage()]);
+            return new UserResource($user);
 
-            return response()->json([
-                'status' => 'error',
-                'message' => 'User registration failed. Please try again later.',
-            ], 500);
+        } catch (\Exception) {
+            return new FailedResource(500, 'An error occurred during login. Please try again later.');
         }
     }
 }
